fix: refactor user store
