--- store/user/getters.js
@@ -1,14 +1,0 @@
-export default {
-  getUser(state) {
-    return state.user
-  },
-  getOrg(state) {
-    return state.organization
-  },
-  getName(state) {
-    return state.user.name
-  },
-  getUserId(state) {
-    return state.user_id
-  }
-}
