--- store/user/state.js
@@ -1,5 +1,0 @@
-export default () => ({
-  user: null,
-  user_id: null,
-  organization: null
-})
